test: quick test rerun

+++ test/standard/testRun.sh
@@ -5,15 +5,21 @@ if [ ! -f "test/shared/shared.sh" ]; then exit 1; fi
 . test/shared/shared.sh
 
 readonly tag="$1"
+remove="$(echo "$@" | grep -qE '(^|\s)no-remove(\s|$)' && echo "false" || echo "false")"
 
 echo "[testRun][INFO] starting tag=$tag"
 docker stop "$TEST_CONTAINER" || true
 docker rm "$TEST_CONTAINER" || true
-if ! docker run -ti --name "$TEST_CONTAINER" --rm -e TEST_MODE=true -e DEBUGGING="${DEBUGGING}" "${REPO}:$tag"; then
+if "$remove"; then
+	if ! docker run -ti --name "$TEST_CONTAINER" --rm -e TEST_MODE=true -e DEBUGGING="${DEBUGGING}" "${REPO}:$tag"; then
+		echo "[testRun][ERROR] run test failed for $tag"
+		exit 1
+	fi
+	docker stop "$TEST_CONTAINER" > /dev/null 2>&1 || true
+	docker rm "$TEST_CONTAINER" > /dev/null 2>&1 || true
+elif ! docker run -ti --name "$TEST_CONTAINER" -e TEST_MODE=true -e DEBUGGING="${DEBUGGING}" "${REPO}:$tag"; then
 	echo "[testRun][ERROR] run test failed for $tag"
 	exit 1
 fi
-docker stop "$TEST_CONTAINER" || true
-docker rm "$TEST_CONTAINER" || true
 
 echo "[testRun][INFO] successful! tag=$tag"
